fix: merge duplicate Surface-Surface Intersection test suites (#1292)
Rebuilt against current main after the Pass 5 monolith split moved these
suites to their own files (the original PR was branched from a pre-split
commit, so it never applied and never triggered CI).
