+++ Tests/OCCTAnalysisTests/SurfaceSurfaceIntersectionTests.swift
@@ -6,11 +6,45 @@ import simd
 
 @Suite("Surface-Surface Intersection")
 struct SurfaceSurfaceIntersectionTests {
-    @Test("Two planes intersect in a line")
-    func twoPlanes() {
+    @Test("Two planes intersect in a line (intersections method)")
+    func twoPlanesIntersectionsMethod() {
         let p1 = Surface.plane(origin: .zero, normal: SIMD3(0, 0, 1))!
         let p2 = Surface.plane(origin: .zero, normal: SIMD3(0, 1, 0))!
         let curves = p1.intersections(with: p2)
+        #expect(curves.count >= 1)
+    }
+
+    @Test("Plane-plane intersection produces line (intersectionCurves method)")
+    func planePlaneIntersection() {
+        // Two planes intersecting at 90 degrees
+        let plane1 = Surface.plane(origin: SIMD3(0, 0, 0), normal: SIMD3(0, 0, 1))!
+        let plane2 = Surface.plane(origin: SIMD3(0, 0, 0), normal: SIMD3(1, 0, 0))!
+        let curves = plane1.intersectionCurves(with: plane2)
+        #expect(curves.count == 1)
+    }
+
+    @Test("Cylinder-plane intersection produces curves")
+    func cylinderPlaneIntersection() {
+        let cylinder = Surface.cylinder(origin: SIMD3(0, 0, 0), axis: SIMD3(0, 0, 1), radius: 5)!
+        let plane = Surface.plane(origin: SIMD3(0, 0, 5), normal: SIMD3(0, 0, 1))!
+        let curves = cylinder.intersectionCurves(with: plane)
+        #expect(curves.count >= 1)
+    }
+
+    @Test("Non-intersecting surfaces produce no curves")
+    func noIntersection() {
+        // Two parallel planes
+        let plane1 = Surface.plane(origin: SIMD3(0, 0, 0), normal: SIMD3(0, 0, 1))!
+        let plane2 = Surface.plane(origin: SIMD3(0, 0, 10), normal: SIMD3(0, 0, 1))!
+        let curves = plane1.intersectionCurves(with: plane2)
+        #expect(curves.isEmpty)
+    }
+
+    @Test("Sphere-plane intersection")
+    func spherePlaneIntersection() {
+        let sphere = Surface.sphere(center: SIMD3(0, 0, 0), radius: 5)!
+        let plane = Surface.plane(origin: SIMD3(0, 0, 3), normal: SIMD3(0, 0, 1))!
+        let curves = sphere.intersectionCurves(with: plane)
         #expect(curves.count >= 1)
     }
 }
